requirements.txt
aiokafka==0.10.0
evidently==0.4.28
jinja2==3.1.6
jupyter==1.1.1
matplotlib==3.10.1
mlflow==2.22.0
numpy==2.2.5
pandas==2.2.3
pyarrow==19.0.1
python-dotenv==1.1.0
PyYAML==6.0.2
scikit-learn==1.6.1
seaborn==0.13.2
tenacity==9.1.2
websockets==15.0.1
joblib==1.5.0
